// src/types/dashboard.ts
// value must be page url
export enum DashboardTab {
  BASIC_DETAILS = "basic-details",
  CONTACT_DETAILS = "contact-details",
  BANK_ACCOUNT_DETAILS = "bank-account-details",
  SCHEME_DETAILS ="scheme-details",
  DEMAT_ACCOUNT_DETAILS = "demat-account-details",
  OTHER_DETAILS = "other-details",
  // DP_DETAILS = "dp-details",
  PHOTO = "photo",
  // SIGNATURE = "signature",
  // IPV_DETAILS = "ipv-details",
  DERIVATIVE = "derivative",
  NOMINEE_DETAILS = "nominee-details",

  E_SIGN_DETAILS = "e-sign-details",
  // SEGMENT_DETAILS = "segment_details",
}